+++ angel-ps/psf/src/main/java/com/tencent/angel/ml/matrix/psf/update/enhance/zip2/Zip2MapWithIndex.java
@@ -46,7 +46,7 @@ public class Zip2MapWithIndex extends MFUpdateFunc {
     DoubleBuffer from2 = rows[1].getData();
     DoubleBuffer to = rows[2].getData();
     int size = rows[0].size();
-    int startCol = rows[0].getStartCol();
+    int startCol = (int)rows[0].getStartCol();
     for (int i = 0; i < size; i++) {
       to.put(i, mapper.call(startCol + i, from1.get(i), from2.get(i)));
     }
